+++ src/WorkOS.net/Services/Organizations/Entities/OrganizationDomain.cs
